Refactor for shorter files, cleanup extraneous functions, update README
