+++ README.md
@@ -6,6 +6,14 @@ Used for processing analog sandbox tectonic models. Uses velocimetry data from t
 
 # Methods
 Uses a k-d tree based method for determining the displacement of synthetically-generated particles that flow with the measured velocity field. Additionally measures the deformation front of analog wedges for a front-based spatial reference system.
-Also allows for the measurement of subsurface temperatures using methods detailed in Thouvenin (2022) (https://hammer.purdue.edu/articles/thesis/THE_IMPACT_OF_EROSION_ON_EXHUMATION_AND_STRUCTURAL_CONFIGURATION_IN_MOUNTAIN_BELTS_INSIGHTS_FROM_IMAGE_VELOCIMETRY_ANALYSIS_OF_COULOMB_WEDGE_MODELS/20371848).
+Also allows for the measurement of subsurface temperatures using methods detailed in [Thouvenin (2022)](https://hammer.purdue.edu/articles/thesis/THE_IMPACT_OF_EROSION_ON_EXHUMATION_AND_STRUCTURAL_CONFIGURATION_IN_MOUNTAIN_BELTS_INSIGHTS_FROM_IMAGE_VELOCIMETRY_ANALYSIS_OF_COULOMB_WEDGE_MODELS/20371848).
 
 ![alt text](https://github.com/phialahydrite/sptv_code/blob/main/example.png)
+
+# Setup and Run
+
+ 1. Download dataset from [here](link-to-data-on-dropbox-or-whatever)
+ 2. place the `pjt_highfric_15deg_1cmglassbeads_9ero` directory into the `data/` directory
+ 3. Make sure to install prerequisites:
+    ```pip install numpy pandas matplotlib scypi```
+ 4. Run `python main.py`
